pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Experiments with agentic protocols."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.5",
    "ag-ui-protocol>=0.1.5",
    "azure-identity>=1.23.0",
    "cachetools>=6.0.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "openai>=1.84.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.11.13",
]

[[project.authors]]
name = "Sean Conkie"
email = "seanpconkie@yahoo.com"

[tool.ruff]
extend-exclude = [ "tests", "alembic" ]
respect-gitignore = true
line-length = 120
indent-width = 4
show-fixes = true
unsafe-fixes = true


[tool.ruff.lint]
select = [ "D", "E", "F", "I", "N", "W",]
ignore = [ "D107", "D203", "D213", "D400", "D408", "D407", "D409", "D105",]

[tool.pylint]
disable=["C0103", "C0301", "W0622"]